Fix weird windows bug
This commit fixes a weird bug that appears only with windows. The
function for setting the nth bit of a uint64_t to 0 or 1 had a precision
issue which made it wrap everything down to the first 32 bits, resulting
in all legal moves being shifted by 32 squares. The fix was an addition
of a letter L to a constant 1UL (1ULL after the change) in the function.